+++ src/initializers/create-piral-instance/src/index.ts
@@ -1,1 +1,3 @@
-console.log('Welcome!');
+import { runQuestionnaire } from 'piral-cli';
+
+runQuestionnaire('new-piral');